inference_api/requirements.txt
uvicorn
Pillow
fastapi
torch
transformers